+++ container-engine-lib/lib/backend_impls/kubernetes/kubernetes_kurtosis_backend_user_service_functions.go
@@ -219,6 +219,7 @@ func (backend *KubernetesKurtosisBackend) StartUserService(
 	serviceGuid service.ServiceGUID,
 	containerImageName string,
 	privatePorts map[string]*port_spec.PortSpec,
+	useStaticPrivatePorts bool, //TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
 	entrypointArgs []string,
 	cmdArgs []string,
 	envVars map[string]string,
@@ -227,6 +228,12 @@ func (backend *KubernetesKurtosisBackend) StartUserService(
 	resultUserService *service.Service,
 	resultErr error,
 ) {
+
+	//TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
+	if useStaticPrivatePorts {
+		logrus.Warn("It's not allowed to use static private ports in Kurtosis Kubernetes Backend, the private port will be automatically generated")
+	}
+
 	preexistingServiceFilters := &service.ServiceFilters{
 		GUIDs:    map[service.ServiceGUID]bool{
 			serviceGuid: true,