+++ container-engine-lib/lib/backend_impls/metrics_reporting/metrics_reporting_kurtosis_backend.go
@@ -323,6 +323,7 @@ func (backend *MetricsReportingKurtosisBackend) StartUserService(
 	guid service.ServiceGUID,
 	containerImageName string,
 	privatePorts map[string]*port_spec.PortSpec,
+	useStaticPrivatePorts bool, //TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
 	entrypointArgs []string,
 	cmdArgs []string,
 	envVars map[string]string,
@@ -337,6 +338,7 @@ func (backend *MetricsReportingKurtosisBackend) StartUserService(
 		guid,
 		containerImageName,
 		privatePorts,
+		useStaticPrivatePorts,
 		entrypointArgs,
 		cmdArgs,
 		envVars,