+++ container-engine-lib/lib/backend_interface/kurtosis_backend.go
@@ -265,6 +265,7 @@ type KurtosisBackend interface {
 		serviceGuid service.ServiceGUID,
 		containerImageName string,
 		privatePorts map[string]*port_spec.PortSpec,
+		useStaticPrivatePorts bool, //TODO this is a huge hack to temporarily enable static ports for NEAR until we have a more productized solution
 		entrypointArgs []string,
 		cmdArgs []string,
 		envVars map[string]string,